Fixed bug with calculating eigenvalues
This bug didn't change the overall result of the program, just the eigenvalues themselves - for comparison purposes.

Also added an option in the info subcommand to view the top eigenvalues (for said comparison purposes).

Fixed a typo in the report, and updated how one of the plots is made (to be more inline with original intentions).